refactored css into scss. Refactored Key Patterns into accordion. Split the types into separate files.
