+++ client/scss/custom.scss
@@ -1,0 +1,15 @@
+// Required
+@import "../node_modules/bootstrap/scss/functions";
+
+// Default variable overrides
+$body-bg: #f9f9f9;
+$body-color: #333;
+$body-line-height: 1.5;
+$body-font-family: 'Montserrat', sans-serif;
+
+// Required
+@import "../node_modules/bootstrap/scss/variables";
+@import "../node_modules/bootstrap/scss/variables-dark";
+@import "../node_modules/bootstrap/scss/maps";
+@import "../node_modules/bootstrap/scss/mixins";
+@import "../node_modules/bootstrap/scss/root";
