+++ client/src/pages/Challenge.tsx
@@ -1,23 +1,50 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 
-import { Container, Row, Col, ButtonToolbar, Button } from "react-bootstrap";
+import { Container, Row, Col, ButtonToolbar, Button, AccordionItem, AccordionHeader, AccordionBody, Accordion } from "react-bootstrap";
 import { useEffect, useState } from "react";
 import { useParams } from "react-router-dom";
-import { ChallengeDetails, ChallengeDifficulties } from "../types";
+import { ChallengeDetails } from "../types/challengeDetails";
+import { ChallengeDifficulties } from "../types/challengeDifficulties";
 import { getDifficulty } from "../utils";
 
-const Challenge = () => {
+//FOR TESTING ONLY
+const challengeDetails: ChallengeDetails = {
+    difficulty: ChallengeDifficulties.HARD,
+    "title": "Intelligent Urban Planning Simulation",
+    "outcome": "A simulation tool for urban planners to model and visualize the impact of development projects on city infrastructure.",
+    "keyPatterns": [
+      "Strategy pattern for different simulation algorithms",
+      "Factory Method pattern for creating various urban model components",
+      "Observer pattern for monitoring changes in the simulation environment",
+      "Mediator pattern for coordinating interactions between simulation components",
+      "Prototype pattern for cloning existing urban models for new simulations"
+    ],
+    "generalDescription": "UrbanSim is a cutting-edge simulation tool that enables urban planners to create detailed models of urban development projects and assess their potential impacts on traffic, population density, public services, and the environment. By offering a variety of components and dynamic interactions, UrbanSim provides a comprehensive view of potential urban transformations.",
+    "expectedFunctionality": {
+      "ModelCreation": "Design comprehensive urban models with diverse components including residential areas, commercial zones, transportation networks, and public services.",
+      "ImpactAnalysis": "Analyze the effects of proposed developments on city infrastructure, such as traffic congestion, environmental sustainability, and public service accessibility.",
+      "SimulationVariants": "Generate and compare different development scenarios to identify optimal urban layouts and policies."
+    },
+    "usageScenarios": {
+      "CloningForExpansionProjects": "Urban planners use UrbanSim to clone the existing urban model as a baseline for proposed expansions. This cloned model is then augmented with new residential areas, commercial centers, and transportation networks. This enables a direct comparison of potential future developments against the current state.",
+      "NewPublicTransportSystem": "Planners simulate the introduction of a new public transport system, evaluating its effects on traffic flow and commuter times. The simulation includes various transport modes and their integration into the existing urban fabric.",
+      "UrbanRenewalProjects": "UrbanSim models the impact of a major urban renewal project, providing insights into potential challenges and benefits for the community. This includes revitalizing old districts, adding green spaces, and improving public amenities.",
+      "PopulationGrowthAdaptation": "Planners use UrbanSim to anticipate changes due to population growth, simulating the expansion of housing, schools, hospitals, and transportation systems to maintain quality of life for residents."
+    }
+  };
 
+const Challenge = () => {
+    //TODO: Fetch this value from the server
     const [completed, setCompleted] = useState(true);
     const [details, setDetails]  = useState<ChallengeDetails>();
     const [showingDetails, setShowingDetails] = useState(false); //toggle state for showing/hiding design pattern list
 
     const { id } = useParams();
 
+
+    //TODO: Refactor once the API is ready
     useEffect(() => {   
         //fetch data about the challenge with the provided "id"
-        //NEEDS TESTING
-        //get the challenge details
         fetch('/api/challenge/' + id).then((response) => {
             if(!response.ok){
                 throw new Error(response.statusText);                
@@ -35,35 +62,10 @@ const Challenge = () => {
 
     //START FOR TESTING ONLY
 
-    const challengeDetails: ChallengeDetails = {
-        difficulty: ChallengeDifficulties.HARD,
-        "title": "Intelligent Urban Planning Simulation",
-        "outcome": "A simulation tool for urban planners to model and visualize the impact of development projects on city infrastructure.",
-        "keyPatterns": [
-          "Strategy pattern for different simulation algorithms",
-          "Factory Method pattern for creating various urban model components",
-          "Observer pattern for monitoring changes in the simulation environment",
-          "Mediator pattern for coordinating interactions between simulation components",
-          "Prototype pattern for cloning existing urban models for new simulations"
-        ],
-        "generalDescription": "UrbanSim is a cutting-edge simulation tool that enables urban planners to create detailed models of urban development projects and assess their potential impacts on traffic, population density, public services, and the environment. By offering a variety of components and dynamic interactions, UrbanSim provides a comprehensive view of potential urban transformations.",
-        "expectedFunctionality": {
-          "ModelCreation": "Design comprehensive urban models with diverse components including residential areas, commercial zones, transportation networks, and public services.",
-          "ImpactAnalysis": "Analyze the effects of proposed developments on city infrastructure, such as traffic congestion, environmental sustainability, and public service accessibility.",
-          "SimulationVariants": "Generate and compare different development scenarios to identify optimal urban layouts and policies."
-        },
-        "usageScenarios": {
-          "CloningForExpansionProjects": "Urban planners use UrbanSim to clone the existing urban model as a baseline for proposed expansions. This cloned model is then augmented with new residential areas, commercial centers, and transportation networks. This enables a direct comparison of potential future developments against the current state.",
-          "NewPublicTransportSystem": "Planners simulate the introduction of a new public transport system, evaluating its effects on traffic flow and commuter times. The simulation includes various transport modes and their integration into the existing urban fabric.",
-          "UrbanRenewalProjects": "UrbanSim models the impact of a major urban renewal project, providing insights into potential challenges and benefits for the community. This includes revitalizing old districts, adding green spaces, and improving public amenities.",
-          "PopulationGrowthAdaptation": "Planners use UrbanSim to anticipate changes due to population growth, simulating the expansion of housing, schools, hospitals, and transportation systems to maintain quality of life for residents."
-        }
-      };
-
     if (details == undefined){
         setDetails(challengeDetails);  
     } 
-    
+
     //END FOR TESTING ONLY
 
     if (details != undefined)
@@ -76,36 +78,31 @@ const Challenge = () => {
                 <h2 className="text-light fw-normal fs-3">{details.outcome}</h2>
             </header> 
         </Row>
+        {completed &&
         <Row>
             {/* If completed, allow to view design patterns. */}
-            {completed && <div className="w-100 mb-4">
-                    {/* <p className="mx-5 text-success"><strong>Completed</strong></p> */}
-                    {/* toggle the state on click and hide/display design pattern list*/}
-                    {!showingDetails ? 
-                        <Button  
-                            variant="outline-success"
-                            onClick={()=>{setShowingDetails(!showingDetails)}}
-                        >View Key Patterns
-                        </Button> : 
-                        <Button 
-                            variant="outline-danger"
-                            onClick={()=>{setShowingDetails(!showingDetails)}}
-                        >Hide Key Patterns</Button>}
-                        <span className="fs-5 float-end text-success">
-                            <strong>Completed</strong>
-                        </span>
-                </div>}
-                <Col xl={10}>
-                {showingDetails && <>
-                    <h3 className="fs-4">Key Patterns</h3>
-                    <ol>    
-                        {details.keyPatterns.map((pattern, index) => {
-                            return <li key={index}>{pattern  + "."}</li>;
-                        })}         
-                    </ol>
-                </>}
+            <Col xl={{span:2, order:2}}> 
+                <span className="fs-5 mb-2 float-end text-success">
+                    <strong>Completed</strong>
+                </span>
             </Col>
-        </Row>
+            <Col xl={{span:10, order: 1}}>
+                <Accordion className="mb-2">
+                    <AccordionItem eventKey="0">
+                        <AccordionHeader className="fs-4">
+                            <h3 className="fs-4">Key Design Patterns</h3>
+                        </AccordionHeader>
+                        <AccordionBody>
+                            <ul>
+                                {details.keyPatterns.map((pattern, index) => {
+                                    return <li key={index}>{pattern + (pattern[-1] != "." && ".")}</li>
+                                })}
+                            </ul>
+                        </AccordionBody>
+                    </AccordionItem>
+                </Accordion>
+            </Col>
+        </Row>}
         <Row >
             <Col xl={10} >
                 <h3 className="fs-4">Description:</h3>
