+++ client/src/types/challengeDifficulties.ts
@@ -1,0 +1,5 @@
+export enum ChallengeDifficulties {
+    EASY,
+    MEDIUM,
+    HARD
+}
