--- client/style.css
@@ -1,6 +1,0 @@
-body {
-    font-family: 'Montserrat', sans-serif !important;
-    line-height: 1.5 !important;
-    color: #333 !important;
-    background-color: #f9f9f9 !important; /* Soft off-white background */
-}
